+++ ui/src/babel-plugin-macros.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+    'fontawesome-svg-core': {
+      'license': 'pro'
+    }
+  }
+
