+++ ui/src/babel.config.js
@@ -1,0 +1,6 @@
+module.exports = function (api) {
+    return {
+      plugins: ['macros'],
+    }
+  }
+
