+++ ui/src/components/sidebar/FileTree.jsx
