+++ ui/src/components/sidebar/Sidebar.jsx
@@ -1,5 +1,5 @@
 import { useContext, useState } from "react"
-import { SlideContext } from '../navbar/SlideContext'
+import { SlideContext } from "../navbar/SlideContext";
 import {GiHamburgerMenu} from "react-icons/gi"
 import FileTree from "./FileTree";
